+++ eras/dijkstra/impl/cddl/lib/Cardano/Ledger/Dijkstra/HuddleSpec.hs
@@ -1018,7 +1018,7 @@ instance HuddleRule "protocol_param_update" DijkstraEra where
         , opt (idx 50 ==> huddleRule @"positive_interval" p) //- "peras healing factor"
         , opt (idx 51 ==> VUInt `sized` (2 :: Word64)) //- "peras certificate boost"
         , opt (idx 52 ==> VUInt `sized` (2 :: Word64)) //- "peras target committee size"
-        , opt (idx 53 ==> (VNil / (VUInt `sized` (8 :: Word64)))) //- "peras bootstrap round"
+        , opt (idx 53 ==> (VNil / (VUInt `sized` (4 :: Word64)))) //- "peras bootstrap round"
         , opt (idx 54 ==> huddleRule @"unit_interval" p) //- "peras quorum threshold safety margin"
         ]
     where